Prohibit negative polynomial term exponents
Because a formal polynomial of x cannot contain either a radical of x or
a rational with x as the denominator, the type of the exponent of
polynomial terms is replaced by int, and asserts that check whether the
exponent is not negative is added to the operations related to
polynomial.

+++ polynomial.h
@@ -9,7 +9,7 @@
 
 typedef struct {
     float coefficient;
-    float exponent;
+    int exponent;
 } term;
 
 typedef struct {
